2018/01: part two working
I've used HashSet instead of HashMap because it is only needed to
check the presence of a key, so there is no values for the keys.

The loop in find_twice() could have a protection against an infinite
loop, but, for the purposes of this exercise, this wasn't needed
because de input is guaranteed (or at least expected) to result in
a second appearance of one frequency.