2018/02: second part
The code for this part is terrible. It really needs more love.

Both the function to calculate the Hamming distance and the function
to get the common letters of a string need a more idiomatic code.
Certainly there is better ways to express them in a more rustic way.

Besides that, I found out that this string comparison is called
Hamming distance and found about the zip() function. :)