2020/01: part two of this day, lots of changes
Added StructOpt so one can set the criteria they want. It's possible to
define the size of combination and the value of the sum.

Change find_pair() to find_combination() for obvious reasons.

Also, of course, there was a better way to access the elements and find
sum and product.