2020/02: even bigger solution for part two
To keep both policies, I had to implement this enormous code structure
to handle both cases.

I changed Entry struct to be less specific. I tried to make Entry<P>
where P is the policy (either the new or the old one), but that was too
hard and I hadn't enough motivation.

I broke the code into modules.

I used FromStr trait both to parse input into Entries and to parse each
line to an Entry.

Structopt saved the day again, making it easy to add arguments to the
application. Some headache to learn that fields in the Opt struct had to
implement FromStr and the FromStr::Err must implement Display too. ()
type doesn't work as it doesn't implement Display...

I'm not happy with the policies.rs file, I think it would be more
elegant if each policy had a type of its own, somehow...